+++ src/tools/kdb/check.hpp
@@ -13,7 +13,7 @@ public:
 
 	virtual std::string getShortOptions()
 	{
-		return "vc";
+		return "vcf";
 	}
 
 	virtual std::string getSynopsis()
@@ -29,24 +29,33 @@ public:
 	virtual std::string getLongHelpText()
 	{
 		return  "If no arguments are given checks on key database\n"
-			"are done instead.\n"
+			"are done instead. Use -f to also do a write test\n"
+			"(might change configuration files!)\n"
 			"\n"
-			"Return values on kdb checking:\n"
+			"For kdb checking a bit pattern will be returned:\n"
 			" 0 .. everything ok (no output)\n"
-			" 1 .. warning on open\n"
-			" 2 .. warning on close\n"
-			" 4 .. error on open (a plugin is broken!)\n"
-			" 8 .. error on close (a plugin is broken!)\n"
+			" bit 1 .. warning on open\n"
+			" bit 2 .. error on open\n"
+			" bit 3 .. warning on get\n"
+			" bit 4 .. error on get\n"
+			" bit 5 .. warning on set (only checked when -f is used)\n"
+			" bit 6 .. error on set (only checked when -f is used)\n"
+			" bit 7 .. warning on close\n"
+			" bit 8 .. error on close\n"
 			"\n"
-			"Or a sum of above\n"
+			"\n"
+			"\n"
+			"\n"
+			"If a plugin name is given, checks will only be done with given plugin.\n"
+			"Use -c to pass options to the plugin.\n"
 			"\n"
 			"Return values on plugin checking:\n"
 			" 0 .. everything ok (no output)\n"
-			" 1 .. no plugin found\n"
+			" 1 .. no such plugin found or plugin could not be opened\n"
 			" 2 .. plugin did not pass checks\n"
 			" 3 .. plugin has warnings\n"
 			"\n"
-			"Please report any issues you found on http://www.libelektra.org\n"
+			"Please report any output caused by official plugins to http://www.libelektra.org\n"
 			"\n";
 	}