+++ doc/NEWS.md
@@ -127,6 +127,7 @@ Raffael Pancheri again did a lot of stabilizing work:
 - Doxygen: remove HTML_TIMESTAMP to make build reproduceable
 - CMake: allow to use qt-gui with qt built with -reduce-relocations
 - fix kdb ls to list warnings during open
+- add -f option to kdb check+improve docu
 
 
 ## Build Server